Put New Profile Screen components on their own files
Resolve https://github.com/shadowsheep1/io-app/pull/2#discussion_r1021405807
